LayoutTests/imported/w3c/web-platform-tests/fetch/nosniff/importscripts-expected.txt


FAIL Test importScripts() assert_equals: expected "PASS" but got "FAIL: Content-Type missing"
